requirements.txt
emoji==2.15.0
fpdf2==2.8.5
googlemaps==4.10.0
langchain_community==0.4.1
langchain_core==1.1.0
langchain_google_genai==3.2.0
langchain_huggingface==1.1.0
langgraph==1.0.4
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.50.0
